requirements.txt
streamlit
httpx
beautifulsoup4
python-dotenv
openai
